requirements.txt
Django==4.2.16
djangorestframework==3.14.0
python-decouple==3.6
drf-spectacular==0.27.2
psycopg2-binary==2.9.7
stripe==7.7.0
twilio==8.12.0
celery==5.3.1
redis==5.0.1
cryptography==41.0.7
hypothesis==6.88.1
pytest==7.4.3
pytest-django==4.7.0
pandas==2.3.1
pydantic==2.11.7
openpyxl==3.1.5
email-validator==2.2.0
geopy==2.4.1
django-cors-headers==4.3.1

# New packages for enterprise features
# Real-time & WebSockets
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Audit logging & change tracking
django-simple-history==3.4.0

# GIS/Mapping support
geopandas==1.0.1
shapely==2.0.6
fiona==1.10.1

# Redistricting & geographic analysis
gerrychain==0.2.22
maup==2.0.1

# AI/NLP for analytics
pydantic-ai==0.0.13
ollama==0.4.4

# Security & MFA
django-otp==1.5.4
django-two-factor-auth==1.17.0
qrcode==8.0

# Impersonation system
django-impersonate==1.9.4

# Push notifications
django-push-notifications==3.1.0
pyfcm==1.5.4
apns2==0.7.2

# Additional utilities
django-flexible-subscriptions==0.16.0
python-magic==0.4.27
pillow==10.4.0